[WIP] Remove ConfigProvider, many unit tests are broken
