+++ ocean_lib/models/test/test_metadata.py
@@ -3,7 +3,6 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-from ocean_lib.config_provider import ConfigProvider
 from ocean_lib.models.metadata import MetadataContract
 from ocean_lib.ocean.test.test_ocean_assets import create_asset
 from ocean_lib.ocean.util import get_contracts_addresses
@@ -11,12 +10,11 @@ from tests.resources.ddo_helpers import wait_for_ddo
 from tests.resources.helper_functions import get_publisher_wallet
 
 
-def test_metadata_contract(publisher_ocean_instance):
+def test_metadata_contract(publisher_ocean_instance, config):
     ocn = publisher_ocean_instance
     alice = get_publisher_wallet()
     block = ocn.web3.eth.blockNumber
 
-    config = ConfigProvider.get_config()
     ddo_address = get_contracts_addresses("ganache", config)[
         MetadataContract.CONTRACT_NAME
     ]
