+++ ocean_lib/ocean/mint_fake_ocean.py
@@ -5,8 +5,6 @@
 import json
 import os
 
-from ocean_lib.config_provider import ConfigProvider
-from ocean_lib.example_config import ExampleConfig
 from ocean_lib.models.data_token import DataToken
 from ocean_lib.ocean.util import get_web3_connection_provider, to_base_18
 from ocean_lib.web3_internal.contract_handler import ContractHandler
@@ -16,14 +14,12 @@ from ocean_lib.web3_internal.wallet import Wallet
 from ocean_lib.web3_internal.web3_provider import Web3Provider
 
 
-def mint_fake_OCEAN():
+def mint_fake_OCEAN(config):
     """
     Does the following:
     1. Mints tokens
     2. Distributes tokens to TEST_PRIVATE_KEY1 and TEST_PRIVATE_KEY2
     """
-    config = ExampleConfig.get_config()
-    ConfigProvider.set_config(config)
     Web3Provider.init_web3(provider=get_web3_connection_provider(config.network_url))
     ContractHandler.set_artifacts_path(config.artifacts_path)
