+++ ocean_lib/ocean/test/test_ocean_services.py
@@ -7,10 +7,10 @@ from ocean_lib.common.utils.utilities import get_timestamp
 from tests.resources.helper_functions import get_publisher_wallet
 
 
-def test_create_access_service(publisher_ocean_instance):
+def test_create_access_service(publisher_ocean_instance, config):
     """Tests that an access service is correctly created."""
     service = publisher_ocean_instance.services.create_access_service(
-        {"a": 1}, "service_endpoint"
+        config, {"a": 1}, "service_endpoint"
     )
     assert service[0] == "access", "Service type is not access."
     assert service[1]["attributes"] == {
@@ -21,7 +21,7 @@ def test_create_access_service(publisher_ocean_instance):
     ), "The service endpoint is different."
 
 
-def test_create_compute_service(publisher_ocean_instance):
+def test_create_compute_service(publisher_ocean_instance, config):
     """Tests that a compute service is correctly created.
 
     Includes cluster, container and server creation."""
@@ -47,7 +47,7 @@ def test_create_compute_service(publisher_ocean_instance):
         3600 * 24, get_publisher_wallet().address, get_timestamp(), provider_attributes
     )
     service = publisher_ocean_instance.services.create_compute_service(
-        attributes, "http://provider.com:8030"
+        config, attributes, "http://provider.com:8030"
     )
     assert (
         isinstance(service, tuple) and len(service) == 2
